Migration tool function
