+++ frontend/components/ui/filePreviewDrawer.tsx
@@ -5,10 +5,31 @@ import { useTranslation } from 'react-i18next';
 import dynamic from 'next/dynamic';
 import { Drawer, Spin, Button, Table } from 'antd';
 import { Download, Maximize2, Minimize2, Minus, Plus, RotateCw, X } from 'lucide-react';
-import Papa from 'papaparse';
 import { FilePreviewProps } from '@/types/chat';
+import { DetectedFileType, ImageBaseMode } from '@/types/file';
+import {
+  CHUNK_SIZE,
+  TEXT_RENDER_BLOCK_SIZE,
+  CSV_ROW_HEIGHT,
+  isValidContainerElement,
+  updateChunkRangeState,
+  ensurePreviewTextDecoder,
+  decodePreviewChunk,
+  decodeLocalTextFile,
+  splitPreviewSafeText,
+  shouldStopFetchingChunk,
+  handlePreviewChunkBoundaryResponse,
+  appendTextPreviewContent,
+  parseCsvLine,
+  detectCsvDelimiter,
+  computeRotateFitScale,
+  clamp,
+  ignoreAbortError,
+  getPageWrapperStyle,
+} from '@/lib/filePreviewUtils';
 import { storageService } from '@/services/storageService';
 import { MarkdownRenderer, extractMarkdownHeadings, type MarkdownHeading } from '@/components/ui/markdownRenderer';
+import { formatFileSize } from '@/lib/utils';
 import log from '@/lib/logger';
 
 const PdfViewer = dynamic(() => import('@/components/ui/PdfViewer').then(mod => ({ default: mod.PdfViewer })), {
@@ -19,355 +40,6 @@ const PdfViewer = dynamic(() => import('@/components/ui/PdfViewer').then(mod => 
     </div>
   ),
 });
-
-const CHUNK_SIZE = 128 * 1024;
-
-const CSV_ROW_HEIGHT = 40;
-const TEXT_RENDER_BLOCK_SIZE = 200;
-const CSV_DELIMITER_CANDIDATES = [',', ';', '\t', '|'] as const;
-const CHARSET_PATTERN = /charset\s*=\s*([^;\s]+)/i;
-const CONTENT_RANGE_PATTERN = /bytes (\d+)-(\d+)\/(\d+)/;
-const INVALID_CONTAINER_TAGS = new Set(['head', 'style', 'script', 'link', 'meta']);
-
-function isValidContainerElement(el: Element | null): el is HTMLDivElement {
-  if (!(el instanceof HTMLDivElement)) {
-    return false;
-  }
-
-  if (!el.isConnected) {
-    return false;
-  }
-
-  const tagName = el.tagName.toLowerCase();
-  return !INVALID_CONTAINER_TAGS.has(tagName);
-}
-
-function normalizeCharsetLabel(value: string): string {
-  const normalized = value.trim().toLowerCase();
-  if (normalized === 'gbk' || normalized === 'gb2312' || normalized === 'cp936') {
-    return 'gb18030';
-  }
-  return normalized;
-}
-
-function extractCharsetFromContentType(contentType: string | null): string | null {
-  if (!contentType) return null;
-  const match = CHARSET_PATTERN.exec(contentType);
-  if (!match?.[1]) return null;
-  return normalizeCharsetLabel(match[1].replaceAll(/^"|"$/g, ''));
-}
-
-function updateChunkRangeState(
-  contentRange: string | null,
-  byteLength: number,
-  byteOffsetRef: React.MutableRefObject<number>,
-  totalBytesRef: React.MutableRefObject<number | null>,
-): boolean {
-  if (!contentRange) {
-    byteOffsetRef.current += byteLength;
-    return false;
-  }
-
-  const match = CONTENT_RANGE_PATTERN.exec(contentRange);
-  if (!match) {
-    byteOffsetRef.current += byteLength;
-    return false;
-  }
-
-  const fetchedEnd = Number(match[2]);
-  const total = Number(match[3]);
-  byteOffsetRef.current = fetchedEnd + 1;
-  totalBytesRef.current = total;
-  return fetchedEnd + 1 < total;
-}
-
-function ensurePreviewTextDecoder(
-  contentType: string | null,
-  textDecoderRef: React.MutableRefObject<TextDecoder | null>,
-  decoderEncodingRef: React.MutableRefObject<string | null>,
-  decoderHasExplicitCharsetRef: React.MutableRefObject<boolean>,
-  decoderAllowGbFallbackRef: React.MutableRefObject<boolean>,
-): void {
-  if (textDecoderRef.current) {
-    return;
-  }
-
-  const headerCharset = extractCharsetFromContentType(contentType);
-  if (headerCharset) {
-    const normalized = normalizeCharsetLabel(headerCharset);
-    const isUtf8 = normalized === 'utf-8' || normalized === 'utf8';
-
-    textDecoderRef.current = isUtf8
-      ? new TextDecoder('utf-8', { fatal: true })
-      : new TextDecoder(normalized);
-    decoderEncodingRef.current = isUtf8 ? 'utf-8' : normalized;
-    decoderHasExplicitCharsetRef.current = true;
-    decoderAllowGbFallbackRef.current = isUtf8;
-    return;
-  }
-
-  // Start with strict UTF-8; if invalid bytes appear in later chunks, fallback to GB18030.
-  textDecoderRef.current = new TextDecoder('utf-8', { fatal: true });
-  decoderEncodingRef.current = 'utf-8';
-  decoderHasExplicitCharsetRef.current = false;
-  decoderAllowGbFallbackRef.current = true;
-}
-
-function decodePreviewChunk(
-  buf: ArrayBuffer,
-  hasMore: boolean,
-  textDecoderRef: React.MutableRefObject<TextDecoder | null>,
-  decoderEncodingRef: React.MutableRefObject<string | null>,
-  decoderAllowGbFallbackRef: React.MutableRefObject<boolean>,
-): string {
-  if (!textDecoderRef.current) {
-    throw new Error('Text decoder is not initialized');
-  }
-
-  try {
-    let raw = textDecoderRef.current.decode(buf, { stream: hasMore });
-    if (!hasMore) {
-      raw += textDecoderRef.current.decode();
-    }
-    return raw;
-  } catch (decodeErr) {
-    const canFallbackToGb18030 =
-      decoderAllowGbFallbackRef.current &&
-      decoderEncodingRef.current === 'utf-8';
-
-    if (!canFallbackToGb18030) {
-      throw decodeErr;
-    }
-
-    log.warn('UTF-8 decode failed for preview stream, fallback to GB18030:', decodeErr);
-    textDecoderRef.current = new TextDecoder('gb18030');
-    decoderEncodingRef.current = 'gb18030';
-    decoderAllowGbFallbackRef.current = false;
-
-    let raw = textDecoderRef.current.decode(buf, { stream: hasMore });
-    if (!hasMore) {
-      raw += textDecoderRef.current.decode();
-    }
-    return raw;
-  }
-}
-
-async function decodeLocalTextFile(file: File): Promise<string> {
-  const buf = await file.arrayBuffer();
-
-  try {
-    return new TextDecoder('utf-8', { fatal: true }).decode(buf);
-  } catch {
-    return new TextDecoder('gb18030').decode(buf);
-  }
-}
-
-function splitPreviewSafeText(
-  raw: string,
-  remainder: string,
-  hasMore: boolean,
-  detectedFileType: DetectedFileType,
-): { remainder: string; safeText: string } {
-  const mergedText = remainder + raw;
-  const shouldKeepTrailingLine = hasMore && detectedFileType !== 'markdown';
-  if (!shouldKeepTrailingLine) {
-    return { remainder: '', safeText: mergedText };
-  }
-
-  const lastNl = mergedText.lastIndexOf('\n');
-  if (lastNl === -1) {
-    return { remainder: mergedText, safeText: '' };
-  }
-
-  return {
-    remainder: mergedText.slice(lastNl + 1),
-    safeText: mergedText.slice(0, lastNl + 1),
-  };
-}
-
-function shouldStopFetchingChunk(
-  activeSessionId: number,
-  currentSessionId: number,
-): boolean {
-  return activeSessionId !== currentSessionId;
-}
-
-function handlePreviewChunkBoundaryResponse(
-  status: number,
-  isFirst: boolean,
-  setServerTooLarge: React.Dispatch<React.SetStateAction<boolean>>,
-  setLoading: React.Dispatch<React.SetStateAction<boolean>>,
-  setLoadingMore: React.Dispatch<React.SetStateAction<boolean>>,
-  observerRef: React.MutableRefObject<IntersectionObserver | null>,
-  isFetchingRef: React.MutableRefObject<boolean>,
-): boolean {
-  if (status === 413) {
-    setServerTooLarge(true);
-    if (isFirst) {
-      setLoading(false);
-    } else {
-      setLoadingMore(false);
-    }
-    isFetchingRef.current = false;
-    return true;
-  }
-
-  if (status === 416) {
-    observerRef.current?.disconnect();
-    if (isFirst) {
-      setLoading(false);
-    } else {
-      setLoadingMore(false);
-    }
-    isFetchingRef.current = false;
-    return true;
-  }
-
-  return false;
-}
-
-function appendTextPreviewContent(
-  params: {
-    detectedFileType: DetectedFileType;
-    safeText: string;
-    byteOffset: number;
-    currentChunkLength: number;
-    csvDelimiterRef: React.MutableRefObject<string>;
-    setTxtLines: React.Dispatch<React.SetStateAction<string[]>>;
-    setCsvRows: React.Dispatch<React.SetStateAction<string[][]>>;
-    setTextContent: React.Dispatch<React.SetStateAction<string>>;
-  },
-): void {
-  const {
-    detectedFileType,
-    safeText,
-    byteOffset,
-    currentChunkLength,
-    csvDelimiterRef,
-    setTxtLines,
-    setCsvRows,
-    setTextContent,
-  } = params;
-
-  if (!safeText) {
-    return;
-  }
-
-  if (detectedFileType === 'text') {
-    const newLines = safeText.split('\n');
-    if (newLines.at(-1) === '') {
-      newLines.pop();
-    }
-    setTxtLines(prev => [...prev, ...newLines]);
-    return;
-  }
-
-  if (detectedFileType === 'csv') {
-    if (byteOffset === currentChunkLength) {
-      csvDelimiterRef.current = detectCsvDelimiter(safeText);
-    }
-    const newLines = safeText.split('\n').filter(line => line.trim().length > 0);
-    setCsvRows(prev => [...prev, ...newLines.map((line) => parseCsvLine(line, csvDelimiterRef.current))]);
-    return;
-  }
-
-  setTextContent(prev => prev + safeText);
-}
-
-function parseCsvLine(line: string, delimiter: string): string[] {
-  const parsed = Papa.parse<string[]>(line, {
-    header: false,
-    skipEmptyLines: false,
-    dynamicTyping: false,
-    delimiter,
-    quoteChar: '"',
-    escapeChar: '"',
-  });
-
-  const row = parsed.data[0];
-  if (Array.isArray(row)) {
-    return row.map((cell) => (typeof cell === 'string' ? cell.trim() : String(cell ?? '').trim()));
-  }
-
-  return line.split(delimiter).map((cell) => cell.trim());
-}
-
-function detectCsvDelimiter(sampleText: string): string {
-  const lines = sampleText
-    .split('\n')
-    .map((line) => line.trim())
-    .filter((line) => line.length > 0)
-    .slice(0, 5);
-
-  if (lines.length === 0) {
-    return ',';
-  }
-
-  let bestDelimiter = ',';
-  let bestScore = -1;
-
-  for (const delimiter of CSV_DELIMITER_CANDIDATES) {
-    const columnCounts = lines.map((line) => {
-      const parsed = Papa.parse<string[]>(line, {
-        header: false,
-        skipEmptyLines: false,
-        dynamicTyping: false,
-        delimiter,
-        quoteChar: '"',
-        escapeChar: '"',
-      });
-
-      const row = parsed.data[0];
-      return Array.isArray(row) ? row.length : 1;
-    });
-
-    const minColumns = Math.min(...columnCounts);
-    const maxColumns = Math.max(...columnCounts);
-    const averageColumns =
-      columnCounts.reduce((sum, count) => sum + count, 0) / columnCounts.length;
-
-    if (averageColumns <= 1) {
-      continue;
-    }
-
-    const consistencyBonus = maxColumns === minColumns ? 100 : 0;
-    const score = consistencyBonus + averageColumns;
-
-    if (score > bestScore) {
-      bestScore = score;
-      bestDelimiter = delimiter;
-    }
-  }
-
-  return bestDelimiter;
-}
-
-function computeRotateFitScale(
-  rotationDeg: number,
-  naturalSize: { width: number; height: number },
-  viewportSize: { width: number; height: number },
-): number {
-  const { width: naturalWidth, height: naturalHeight } = naturalSize;
-  const { width: viewportWidth, height: viewportHeight } = viewportSize;
-  if (naturalWidth <= 0 || naturalHeight <= 0 || viewportWidth <= 0 || viewportHeight <= 0) {
-    return 1;
-  }
-
-  const normalizedRotation = ((rotationDeg % 360) + 360) % 360;
-  const isQuarterTurn = normalizedRotation === 90 || normalizedRotation === 270;
-  const rotatedWidth = isQuarterTurn ? naturalHeight : naturalWidth;
-  const rotatedHeight = isQuarterTurn ? naturalWidth : naturalHeight;
-  const fitScale = Math.min(viewportWidth / rotatedWidth, viewportHeight / rotatedHeight);
-  return Number.isFinite(fitScale) && fitScale > 0 ? fitScale : 1;
-}
-
-function clamp(value: number, min: number, max: number): number {
-  return Math.min(Math.max(value, min), max);
-}
-
-type ImageBaseMode = 'fit' | 'actual';
-
-type DetectedFileType = 'pdf' | 'image' | 'markdown' | 'csv' | 'text' | 'office' | 'unknown';
 
 export function FilePreviewDrawer(props: Readonly<FilePreviewProps>) {
   const { open, onClose } = props;
@@ -691,11 +363,14 @@ export function FilePreviewDrawer(props: Readonly<FilePreviewProps>) {
   }, []);
 
   const handleImageDoubleClick = useCallback(() => {
-    setImageBaseMode('fit');
-    setImageScale(1);
-    imageScaleRef.current = 1;
-    handleImagePanReset();
-  }, [handleImagePanReset]);
+    if (imageScale !== 1 || imageBaseMode !== 'fit') {
+      setImageBaseMode('fit');
+      setImageScale(1);
+      imageScaleRef.current = 1;
+    } else {
+      setImageBaseMode('actual');
+    }
+  }, [imageBaseMode, imageScale]);
 
   const toggleImageBaseMode = useCallback(() => {
     if (imageBaseMode === 'fit') {
@@ -1033,14 +708,6 @@ export function FilePreviewDrawer(props: Readonly<FilePreviewProps>) {
       setShowMarkdownToc(false);
     }
   }, []);
-
-  const formatFileSize = (size: number): string => {
-    if (size < 1024) return `${size} B`;
-    if (size < 1024 * 1024) return `${(size / 1024).toFixed(1)} KB`;
-    return `${(size / (1024 * 1024)).toFixed(2)} MB`;
-  };
-
-
 
   const renderLoading = () => (
     <div className="flex items-center justify-center h-full">
